+++ core/src/com/unciv/models/ruleset/unique/Unique.kt
@@ -52,7 +52,7 @@ class Unique(val text: String, val sourceObjectType: UniqueTarget? = null, val s
     fun hasModifier(type: UniqueType) = modifiersMap.containsKey(type)
     fun isModifiedByGameSpeed() = hasModifier(UniqueType.ModifiedByGameSpeed)
     fun isModifiedByGameProgress() = hasModifier(UniqueType.ModifiedByGameProgress)
-    fun modifyByGameProgress(civ: Civilization, modifierPercent: Int): Float {
+    fun modifyByGameProgress(civ: Civilization, modifierPercent: Float): Float {
         // Civ6 game progression modifier: https://www.reddit.com/r/civ/comments/gvx44v/comment/fsrifc2/
         var modifier = 1f
         val ruleset = civ.gameInfo.ruleset
